+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## v1.2.1
+
+### Added
+- Added function to create a support bundle
+- Added function to download a support bundle
+- Added function to list support bundles
+- Added function to delete a support bundle
+
+
+## v1.1.1
+
+### Added
+- Added function to create local repo
+
+### Updated
+- Bug fixes
+
 ## v1.0.0
 
 ### Added